requirements.txt
colorama
speechRecognition
pyttsx3
pyaudio
requests
google-generativeai
